pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-code-reviewer"
version = "0.1.0"
description = "Smart Code Reviewer: CLI + local web UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.0",
  "httpx>=0.27.0",
  "eval_type_backport>=0.2.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0"
]

[project.scripts]
scr = "scr.cli:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
